requirements.txt
Flask>=2.3.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
reportlab>=4.0.4
openpyxl>=3.1.0
flask-cors>=4.0.0
Werkzeug>=2.3.0
